[S09P22B310-000] feat : 멤버 역할 변경 api 추가

+++ backend/src/main/java/com/ntt/mwonimoney/domain/member/entity/Parent.java
@@ -16,7 +16,7 @@ import lombok.NoArgsConstructor;
 public class Parent extends Member {
 
 	@Builder
-	public Parent(byte status, String name, String nickname, String birthday, SocialProvider socialProvider,
+	public Parent(int status, String name, String nickname, String birthday, SocialProvider socialProvider,
 		String socialId, String email) {
 		super(status, name, nickname, birthday, socialProvider, socialId, email, MemberRole.PARENT);
 	}
